Feat : 신규 약 등록 기능 구현

+++ src/main/java/com/pioneer/medgo/domain/StockDTO.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 public class StockDTO {
 	public Long medicineId;
 	public Long pharmacyId;
-	public Long medCount;
+	public int medCount;
 	public String productName;
 	public String manufacturerName;
 	public String mainCode;
